requirements.txt
Django==4.0.3
gunicorn==20.1.0
djangorestframework==3.13.1
markdown==3.3.6
django-filter==21.1
Pillow==9.1.0
psycopg2-binary==2.9.3
django-ckeditor==6.4.0